pyproject.toml
[tool.poetry]
name = "huemid"
version = "0.1.0"
description = ""
authors = ["rndmit <rndmit@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anticaptchaofficial = "^1.0.45"
selenium = "^4.6.0"
ksuid = "^1.3"
python-telegram-bot = "^13.14"



[tool.poetry.group.dev.dependencies]
pylama = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
